moved tags subcommands to own classes & added tag name editing
